Separate postgres-related backup from general backup.go
