+++ internal/databases/sqlserver/backup_push_handler.go
@@ -7,6 +7,8 @@ import (
 	"os"
 	"syscall"
 
+	"github.com/wal-g/wal-g/internal/databases/postgres"
+
 	"github.com/wal-g/tracelog"
 	"github.com/wal-g/wal-g/internal"
 	"github.com/wal-g/wal-g/internal/databases/sqlserver/blob"
@@ -48,7 +50,7 @@ func HandleBackupPush(dbnames []string, updateLatest bool, compression bool) {
 		tracelog.ErrorLogger.FatalfOnError("can't find latest backup: %v", err)
 		backupName = backup.Name
 		sentinel = new(SentinelDto)
-		err = internal.FetchStreamSentinel(backup, sentinel)
+		err = backup.FetchSentinel(&sentinel)
 		tracelog.ErrorLogger.FatalOnError(err)
 		sentinel.Databases = uniq(append(sentinel.Databases, dbnames...))
 	} else {
@@ -67,7 +69,7 @@ func HandleBackupPush(dbnames []string, updateLatest bool, compression bool) {
 	sentinel.StopLocalTime = utility.TimeNowCrossPlatformLocal()
 	uploader := internal.NewUploader(nil, folder.GetSubFolder(utility.BaseBackupPath))
 	tracelog.InfoLogger.Printf("uploading sentinel: %s", sentinel)
-	err = internal.UploadSentinel(uploader, sentinel, backupName)
+	err = postgres.UploadSentinel(uploader, sentinel, backupName)
 	tracelog.ErrorLogger.FatalfOnError("failed to save sentinel: %v", err)
 
 	tracelog.InfoLogger.Printf("backup finished")
