+++ internal/stream_fetch_helper.go
@@ -1,13 +1,10 @@
 package internal
 
 import (
-	"encoding/json"
 	"fmt"
 	"io"
 	"time"
 
-	"github.com/pkg/errors"
-	"github.com/wal-g/storages/storage"
 	"github.com/wal-g/wal-g/internal/compression"
 	"github.com/wal-g/wal-g/utility"
 )
@@ -36,13 +33,12 @@ func GetLogsDstSettings(operationLogsDstEnvVariable string) (dstFolder string, e
 
 // TODO : unit tests
 // downloadAndDecompressStream downloads, decompresses and writes stream to stdout
-func downloadAndDecompressStream(backup *Backup, writeCloser io.WriteCloser) error {
+func downloadAndDecompressStream(backup Backup, writeCloser io.WriteCloser) error {
 	defer writeCloser.Close()
 
 	for _, decompressor := range compression.Decompressors {
-		archiveReader,
-			exists,
-			err := TryDownloadFile(backup.BaseBackupFolder, GetStreamName(backup.Name, decompressor.FileExtension()))
+		archiveReader, exists, err := TryDownloadFile(
+			backup.Folder, GetStreamName(backup.Name, decompressor.FileExtension()))
 		if err != nil {
 			return err
 		}
@@ -58,36 +54,4 @@ func downloadAndDecompressStream(backup *Backup, writeCloser io.WriteCloser) err
 		return nil
 	}
 	return newArchiveNonExistenceError(fmt.Sprintf("Archive '%s' does not exist.\n", backup.Name))
-}
-
-// TODO : unit tests
-func FetchStreamSentinel(backup *Backup, sentinelDto interface{}) error {
-	sentinelDtoData, err := backup.fetchSentinelData()
-	if err != nil {
-		return errors.Wrap(err, "failed to fetch sentinel")
-	}
-	err = json.Unmarshal(sentinelDtoData, sentinelDto)
-	return errors.Wrap(err, "failed to unmarshal sentinel")
-}
-
-// DownloadFile downloads, decompresses and decrypts
-func DownloadFile(folder storage.Folder, filename, ext string, writeCloser io.WriteCloser) error {
-	decompressor := compression.FindDecompressor(ext)
-	if decompressor == nil {
-		return fmt.Errorf("decompressor for extension '%s' was not found", ext)
-	}
-	archiveReader, exists, err := TryDownloadFile(folder, filename)
-	if err != nil {
-		return err
-	}
-	if !exists {
-		return fmt.Errorf("file '%s' does not exist", filename)
-	}
-
-	err = DecompressDecryptBytes(&EmptyWriteIgnorer{WriteCloser: writeCloser}, archiveReader, decompressor)
-	if err != nil {
-		return err
-	}
-	utility.LoggedClose(writeCloser, "")
-	return nil
 }
